# Optimize cast (timestamp as string)