--- a/velox/type/Timestamp.cpp
+++ b/velox/type/Timestamp.cpp
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 #include "velox/type/Timestamp.h"
+#include <charconv>
 #include <chrono>
+#include "velox/common/base/CountBits.h"
 #include "velox/external/date/tz.h"
 #include "velox/type/tz/TimeZoneMap.h"
 
@@ -185,23 +187,6 @@ void appendSmallInt(int n, std::string& out) {
   VELOX_DCHECK_LE(n, 61);
   out.append(intToStr[n], 2);
 }
-
-std::string::size_type getCapacity(const TimestampToStringOptions& options) {
-  auto precisionWidth = static_cast<int8_t>(options.precision);
-  switch (options.mode) {
-    case TimestampToStringOptions::Mode::kDateOnly:
-      // yyyy-mm-dd
-      return 10;
-    case TimestampToStringOptions::Mode::kTimeOnly:
-      // hh:mm:ss.precision
-      return 9 + precisionWidth;
-    case TimestampToStringOptions::Mode::kFull:
-      return 26 + precisionWidth;
-    default:
-      VELOX_UNREACHABLE();
-  }
-}
-
 } // namespace
 
 bool Timestamp::epochToUtc(int64_t epoch, std::tm& tm) {
@@ -249,54 +234,81 @@ bool Timestamp::epochToUtc(int64_t epoch, std::tm& tm) {
   return true;
 }
 
-std::string Timestamp::tmToString(
+StringView Timestamp::tmToStringView(
     const std::tm& tmValue,
     uint64_t nanos,
-    const TimestampToStringOptions& options) {
+    const TimestampToStringOptions& options,
+    char* const startPosition) {
   VELOX_DCHECK_GE(nanos, 0);
   VELOX_DCHECK_LT(nanos, 1'000'000'000);
-  auto precisionWidth = static_cast<int8_t>(options.precision);
-  std::string out;
-  out.reserve(getCapacity(options));
 
+  const auto appendDigits = [](const int value,
+                               const std::optional<uint32_t> minWidth,
+                               char* const position) {
+    const auto numDigits = countDigits(value);
+    uint32_t offset = 0;
+    // Append leading zeros when there is the requirement for minumum width.
+    if (minWidth.has_value() && numDigits < minWidth.value()) {
+      const auto leadingZeros = minWidth.value() - numDigits;
+      std::memset(position, '0', leadingZeros);
+      offset += leadingZeros;
+    }
+    const auto [endPosition, errorCode] =
+        std::to_chars(position + offset, position + offset + numDigits, value);
+    VELOX_DCHECK_EQ(
+        errorCode,
+        std::errc(),
+        "Failed to convert value to varchar: {}.",
+        std::make_error_code(errorCode).message());
+    offset += numDigits;
+    return offset;
+  };
+
+  char* writePosition = startPosition;
   if (options.mode != TimestampToStringOptions::Mode::kTimeOnly) {
-    int n = kTmYearBase + tmValue.tm_year;
-    const bool leadingPositiveSign = options.leadingPositiveSign && n > 9999;
-    bool negative = n < 0;
+    int year = kTmYearBase + tmValue.tm_year;
+    const bool leadingPositiveSign = options.leadingPositiveSign && year > 9999;
+    const bool negative = year < 0;
+
+    // Sign.
     if (negative) {
-      out += '-';
-      n = -n;
-    }
-    while (n > 0) {
-      out += '0' + n % 10;
-      n /= 10;
-    }
-    auto zeroPaddingYearSize = negative + 4;
-    if (options.zeroPaddingYear && out.size() < zeroPaddingYearSize) {
-      while (out.size() < zeroPaddingYearSize) {
-        out += '0';
-      }
-    }
-    if (leadingPositiveSign) {
-      out += '+';
-    }
-    std::reverse(out.begin() + negative, out.end());
-    out += '-';
-    appendSmallInt(1 + tmValue.tm_mon, out);
-    out += '-';
-    appendSmallInt(tmValue.tm_mday, out);
-    if (options.mode == TimestampToStringOptions::Mode::kDateOnly) {
-      return out;
+      *writePosition++ = '-';
+      year = -year;
+    } else if (leadingPositiveSign) {
+      *writePosition++ = '+';
     }
 
-    out += options.dateTimeSeparator;
+    // Year.
+    writePosition += appendDigits(
+        year,
+        options.zeroPaddingYear ? std::optional<uint32_t>(4) : std::nullopt,
+        writePosition);
+
+    // Month.
+    *writePosition++ = '-';
+    writePosition += appendDigits(1 + tmValue.tm_mon, 2, writePosition);
+
+    // Day.
+    *writePosition++ = '-';
+    writePosition += appendDigits(tmValue.tm_mday, 2, writePosition);
+
+    if (options.mode == TimestampToStringOptions::Mode::kDateOnly) {
+      return StringView(startPosition, writePosition - startPosition);
+    }
+    *writePosition++ = options.dateTimeSeparator;
   }
 
-  appendSmallInt(tmValue.tm_hour, out);
-  out += ':';
-  appendSmallInt(tmValue.tm_min, out);
-  out += ':';
-  appendSmallInt(tmValue.tm_sec, out);
+  // Hour.
+  writePosition += appendDigits(tmValue.tm_hour, 2, writePosition);
+
+  // Minute.
+  *writePosition++ = ':';
+  writePosition += appendDigits(tmValue.tm_min, 2, writePosition);
+
+  // Second.
+  *writePosition++ = ':';
+  writePosition += appendDigits(tmValue.tm_sec, 2, writePosition);
+
   if (options.precision == TimestampToStringOptions::Precision::kMilliseconds) {
     nanos /= 1'000'000;
   } else if (
@@ -304,33 +316,77 @@ std::string Timestamp::tmToString(
     nanos /= 1'000;
   }
   if (options.skipTrailingZeros && nanos == 0) {
-    return out;
+    return StringView(startPosition, writePosition - startPosition);
   }
-  out += '.';
-  const int offset = out.size();
-  int trailingZeros = 0;
 
+  // Fractional part.
+  *writePosition++ = '.';
+  // Append leading zeros.
+  const auto numDigits = countDigits(nanos);
+  const auto precisionWidth = static_cast<int8_t>(options.precision);
+  std::memset(writePosition, '0', precisionWidth - numDigits);
+  writePosition += precisionWidth - numDigits;
+
+  // Append the remaining numeric digits.
   if (options.skipTrailingZeros) {
+    std::optional<uint32_t> nonZeroOffset = std::nullopt;
+    int32_t offset = numDigits - 1;
+    // Write non-zero digits from end to start.
     while (nanos > 0) {
-      if (out.size() == offset && nanos % 10 == 0) {
-        trailingZeros += 1;
-      } else {
-        out += '0' + nanos % 10;
+      if (nonZeroOffset.has_value() || nanos % 10 != 0) {
+        *(writePosition + offset) = '0' + nanos % 10;
+        if (!nonZeroOffset.has_value()) {
+          nonZeroOffset = offset;
+        }
       }
+      --offset;
       nanos /= 10;
     }
+    writePosition += nonZeroOffset.value() + 1;
   } else {
-    while (nanos > 0) {
-      out += '0' + nanos % 10;
-      nanos /= 10;
-    }
+    const auto [position, errorCode] =
+        std::to_chars(writePosition, writePosition + numDigits, nanos);
+    VELOX_DCHECK_EQ(
+        errorCode,
+        std::errc(),
+        "Failed to convert fractional part to chars: {}.",
+        std::make_error_code(errorCode).message());
+    writePosition = position;
   }
+  return StringView(startPosition, writePosition - startPosition);
+}
 
-  while (out.size() - offset < precisionWidth - trailingZeros) {
-    out += '0';
+StringView Timestamp::tsToStringView(
+    const Timestamp& ts,
+    const TimestampToStringOptions& options,
+    char* const startPosition) {
+  std::tm tmValue;
+  VELOX_USER_CHECK(
+      epochToUtc(ts.getSeconds(), tmValue),
+      "Can't convert seconds to time: {}",
+      ts.getSeconds());
+  const uint64_t nanos = ts.getNanos();
+  return tmToStringView(tmValue, nanos, options, startPosition);
+}
+
+std::string::size_type getMaxStringLength(
+    const TimestampToStringOptions& options) {
+  const auto precisionWidth = static_cast<int8_t>(options.precision);
+  switch (options.mode) {
+    case TimestampToStringOptions::Mode::kDateOnly:
+      // Date format is %y-mm-dd, where y has 10 digits at maximum for int32.
+      // Possible sign is considered.
+      return 17;
+    case TimestampToStringOptions::Mode::kTimeOnly:
+      // hh:mm:ss.precision
+      return 9 + precisionWidth;
+    case TimestampToStringOptions::Mode::kFull:
+      // Timestamp format is %y-%m-%dT%h:%m:%s.precision, where y has 10 digits
+      // at maximum for int32. Possible sign is considered.
+      return 27 + precisionWidth;
+    default:
+      VELOX_UNREACHABLE();
   }
-  std::reverse(out.begin() + offset, out.end());
-  return out;
 }
 
 void parseTo(folly::StringPiece in, ::facebook::velox::Timestamp& out) {