Previously, cast (timestamp as string) relies on `tmToString` method, which 
creates intermediate strings during conversion. This PR implements
`tsToStringView` and `tmToStringView` which write chars to the provided
position directly. With them, cast (timestamp as string) can be optimized by
avoiding intermediate strings. Below benchmark shows the performance is
improved from 58.14ms to 32.79ms.

| cast##cast_timestamp  | time/iter | iters/s |
| ------------- | ------------- | ------------- |
| before | 58.14ms  |  17.20  | 
| after  | 32.79ms  | 30.50  |



Fixes https://github.com/facebookincubator/velox/issues/8067.