@Yuhta 2024-03-12
> after | 52.40 | 19.09

@rui-mo Are you sure these are under `-O3`?  It was 14.22ms and 17.89ms in master now