@rui-mo 2024-03-13
> @rui-mo Are you sure these are under `-O3`? It was 14.22ms in #6801 and 17.89ms in master now

@Yuhta No, these are not under `-O3`. The benchmark was compiled as debug type, and from `_build/debug/` folder. I tried `-O3` in release type, and got `27.64ms` on master and `9.19ms` with this change. Thanks.