@mbasmanova 2024-03-13
> The benchmark was compiled as debug type,

We should never run benchmarks in debug mode. Am I missing something?