@rui-mo 2024-03-13
> We should never run benchmarks in debug mode. Am I missing something?

@mbasmanova @Yuhta Sorry, it's my mistake. Updated the PR description with the new performance.